backend/requirement.txt
openai
requests
docker
nmap
scapy
python-dotenv
